A shitty title, but the content is worth it

+++ graphics.cpp
@@ -4,7 +4,7 @@ void Graphics::begin() {
    lcd.begin(84, 48);
 }
 
-void Graphics::set_cell(int row, int col, int new_bit) {
+void Graphics::set_cell(byte row, byte col, byte new_bit) {
   if(row >= 48 || col >= 84)
     return;
     
@@ -18,7 +18,7 @@ void Graphics::set_cell(int row, int col, int new_bit) {
       data[byte_index] &= ~bit_mask; // off
 }
  
-int Graphics::get_cell(int row, int col) {
+int Graphics::get_cell(byte row, byte col) {
    int byte_index = col + 84*(row/8);
    int bit_index = row % 8;
    int bit_mask = 1 << bit_index;
@@ -35,7 +35,7 @@ void Graphics::clear() {
   lcd.clear();
 }
 
-void Graphics::text(const char *str, int row, int col) {
+void Graphics::text(const char *str, byte row, byte col) {
   lcd.setCursor(col, row);
   lcd.print(str);
 }
